plex-setup/requirements.txt
plexapi
PyYaml
